Extract runtime
