+++ runtime/src/update_did_signed_extension.rs
@@ -1,6 +1,5 @@
-use crate::{balances, identity, identity::LinkedKeyInfo, Runtime};
-
-use primitives::{AccountId, IdentityId, Key, TransactionError};
+use crate::{balances, identity, identity::LinkedKeyInfo, runtime, Runtime};
+use primitives::{IdentityId, Key, TransactionError};
 
 use sr_primitives::{
     traits::SignedExtension,
@@ -11,11 +10,12 @@ use sr_primitives::{
 
 use codec::{Decode, Encode};
 use core::convert::TryFrom;
+use rstd::marker::PhantomData;
 use srml_support::dispatch::DispatchInfo;
 
 type Identity = identity::Module<Runtime>;
 type Balance = balances::Module<Runtime>;
-type Call = crate::Call;
+type Call = runtime::Call;
 
 /// This signed extension double-checks and updates the current identifier extracted from
 /// the caller account in each transaction.
@@ -24,12 +24,15 @@ type Call = crate::Call;
 /// - After transaction, Do we need to clean `CurrentDid`?
 /// - Move outside `lib.rs`. It needs a small refactor to extract `Runtime` from here.
 #[derive(Default, Encode, Decode, Clone, Eq, PartialEq)]
-#[cfg_attr(feature = "std", derive(Debug))]
-pub struct UpdateDid;
+pub struct UpdateDid<T: system::Trait + Send + Sync>(PhantomData<T>);
 
-impl UpdateDid {
+impl<T: system::Trait + Send + Sync> UpdateDid<T> {
+    fn new() -> Self {
+        UpdateDid(PhantomData)
+    }
+
     /// It extracts the `IdentityId` associated with `who` account.
-    fn identity_from_key(who: &AccountId) -> Option<IdentityId> {
+    fn identity_from_key(who: &T::AccountId) -> Option<IdentityId> {
         if let Ok(who_key) = Key::try_from(who.encode()) {
             if let Some(linked_key_info) = Identity::key_to_identity_ids(&who_key) {
                 if let LinkedKeyInfo::Unique(id) = linked_key_info {
@@ -41,9 +44,17 @@ impl UpdateDid {
     }
 }
 
-impl SignedExtension for UpdateDid {
-    type AccountId = AccountId;
-    type Call = Call;
+#[cfg(feature = "std")]
+impl<T: system::Trait + Send + Sync> rstd::fmt::Debug for UpdateDid<T> {
+    fn fmt(&self, f: &mut rstd::fmt::Formatter) -> rstd::fmt::Result {
+        write!(f, "UpdateDid")
+    }
+}
+
+impl<T: system::Trait + Send + Sync> SignedExtension for UpdateDid<T> {
+    type AccountId = T::AccountId;
+    // type AccountId = AccId;
+    type Call = runtime::Call;
     type AdditionalSigned = ();
     type Pre = ();
 
@@ -90,17 +101,29 @@ impl SignedExtension for UpdateDid {
 mod tests {
     use super::UpdateDid;
     use crate::{
-        balances, identity,
+        balances, identity, runtime,
         test::storage::{build_ext, make_account_with_balance, TestStorage},
         Runtime,
     };
+    use codec::Encode;
+    use core::default::Default;
     use primitives::TransactionError;
     use sr_io::with_externalities;
-    use sr_primitives::transaction_validity::{InvalidTransaction, ValidTransaction};
+    use sr_primitives::{
+        traits::{SignedExtension, Verify},
+        transaction_validity::{InvalidTransaction, ValidTransaction},
+        AnySignature,
+    };
     use srml_support::dispatch::DispatchInfo;
+    use test_client::{self, AccountKeyring};
+
+    type AccountId = <AnySignature as Verify>::Signer;
 
     type Identity = identity::Module<TestStorage>;
-    type Balance = balances::Module<TestStorage>;
+    type Balances = balances::Module<TestStorage>;
+    type Call = runtime::Call;
+    type BalancesCall = balances::Call<Runtime>;
+    type IdentityCall = identity::Call<Runtime>;
 
     #[test]
     fn update_did_tests() {
@@ -108,23 +131,23 @@ mod tests {
     }
 
     fn update_did_tests_with_externalities() {
-        let update_did_se = UpdateDid::default();
+        let update_did_se = <UpdateDid<TestStorage>>::new();
         let dispatch_info = DispatchInfo::default();
 
-        let (a_acc, b_acc, c_acc, d_acc) = (1u64, 2u64, 3u64, 4u64);
+        let (a_acc, b_acc, c_acc) = (1u64, 2u64, 3u64);
         let (alice, alice_id) = make_account_with_balance(a_acc, 10_000).unwrap();
-        let (bob, bob_id) = make_account_with_balance(b_acc, 5_000).unwrap();
+        let (bob, _bob_id) = make_account_with_balance(b_acc, 5_000).unwrap();
 
-        /*
-        let register_did_call_1 = Runtime::Call::Identity( Identity::Call::register_did( alice_id.clone(), vec![]));
+        let register_did_call_1 = Call::Identity(IdentityCall::register_did(alice_id, vec![]));
         assert_eq!(
-            update_did_se.validate( &a_acc, register_did_call_1, dispatch_info, 0usize),
-            Ok(ValidTransaction::default()));
+            update_did_se.validate(&a_acc, &register_did_call_1, dispatch_info, 0usize),
+            Ok(ValidTransaction::default())
+        );
 
-        let transfer_call_1 = Runtime::Call::Balance( Balance::Call::transfer( b_acc, a_acc, 100));
+        let freeze_call1 = Call::Identity(IdentityCall::freeze_signing_keys(alice_id));
         assert_eq!(
-            update_did_se.validate( &b_acc, transfer_call_1, dispatch_info, 0usize),
-            Err( InvalidTransaction::Custom(TransactionError::MissingIdentity as u8).into()));
-        */
+            update_did_se.validate(&c_acc, &freeze_call1, dispatch_info, 0usize),
+            Err(InvalidTransaction::Custom(TransactionError::MissingIdentity as u8).into())
+        );
     }
 }
